Add android support for cloud

+++ shared-ui/src/commonMain/kotlin/com/github/enteraname74/soulsearching/commondelegate/PlaylistBottomSheetDelegate.kt
@@ -60,8 +60,10 @@ class PlaylistBottomSheetDelegateImpl(
                     setBottomSheetState(null)
 
                     loadingManager.withLoadingOnIO {
-                        val result = deletePlaylistUseCase(playlistToDelete)
-                        feedbackPopUpManager.showResultErrorIfAny(result)
+                        if (!playlistToDelete.isFavorite) {
+                            val result = deletePlaylistUseCase(playlistToDelete)
+                            feedbackPopUpManager.showResultErrorIfAny(result)
+                        }
                         multiSelectionManagerImpl?.clearMultiSelection()
                     }
                 },
